+++ jabkit/src/main/java/org/jabref/cli/Fetch.java
@@ -1,6 +1,20 @@
 package org.jabref.cli;
 
+import java.util.List;
+import java.util.Optional;
+import java.util.Set;
 import java.util.concurrent.Callable;
+
+import org.jabref.logic.importer.FetcherException;
+import org.jabref.logic.importer.ParserResult;
+import org.jabref.logic.importer.SearchBasedFetcher;
+import org.jabref.logic.importer.WebFetchers;
+import org.jabref.logic.l10n.Localization;
+import org.jabref.model.entry.BibEntry;
+
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import picocli.CommandLine;
 
 import static picocli.CommandLine.Command;
 import static picocli.CommandLine.Option;
@@ -13,24 +27,31 @@ import static picocli.CommandLine.Option;
  */
 @Command(name = "fetch", description = "Fetch entries from a provider.")
 class Fetch implements Callable<Integer> {
+    private static final Logger LOGGER = LoggerFactory.getLogger(ArgumentProcessor.class);
+
+    @CommandLine.ParentCommand
+    private KitCommandLine kitCommandLine;
+
     @Option(names = "--provider", required = true)
-    String provider;
+    private String provider;
 
     @Option(names = "--query")
-    String query;
+    private String query;
 
     @Option(names = "--output", required = true)
-    String output;
+    private String output;
 
     @Option(names = "--append", description = "Append to existing file")
-    boolean append;
+    private boolean append;
 
     @Override
     public Integer call() {
-/*        if ((fetchCommand == null) || !fetchCommand.contains(":")) {
+        if ((provider == null) || !provider.contains(":")) {
             System.out.println(Localization.lang("Expected syntax for --fetch='<name of fetcher>:<query>'"));
             System.out.println(Localization.lang("The following fetchers are available:"));
-            return Optional.empty();
+            System.out.println(kitCommandLine.)
+
+            return 1;
         }
 
         String[] split = fetchCommand.split(":");
@@ -67,7 +88,7 @@ class Fetch implements Callable<Integer> {
                 LOGGER.error("Error while fetching", e);
                 return Optional.empty();
             }
-        } */
+        }
 
         System.out.println("Not yet implemented");
         return 0;
